Split utils and constants to allow tests
