+++ packages/subgraph/src/controller.ts
@@ -27,20 +27,12 @@ import {
   HourStatSnapshot,
   DayStatSnapshot,
   VaultHistory,
-  Account,
 } from "../generated/schema";
-import {
-  buyOrSellETH,
-  createTransactionHistory,
-  loadOrCreateAccount,
-} from "./util";
+import { buyOrSellETH, createTransactionHistory } from "./util";
 
-import {
-  BIGINT_ONE,
-  BIGINT_ZERO,
-  SHORT_HELPER_ADDR,
-  EMPTY_ADDR,
-} from "./constants";
+import { BIGINT_ONE, BIGINT_ZERO } from "./constants";
+import { loadOrCreateAccount } from "./utils/loadInit";
+import { EMPTY_ADDR, SHORT_HELPER_ADDR } from "./addresses";
 
 // Note: If a handler doesn't require existing field values, it is faster
 // _not_ to load the entity from the store. Instead, create it fresh with
